Isolate DAG map and loop execution strategies
Map and loop node execution were embedded in the DAG executor core, which made the orchestration loop harder to audit before deeper workflow refactors. The strategy modules keep child-graph recursion, lifecycle hooks, and scoped map concurrency explicit while preserving the public DAGExecutor contract.

Constraint: Issue #1984 calls for extracting loop/map node execution strategies after locking behavior with focused tests.
Rejected: Extract every composite node strategy at once | branch, parallel, and sub-workflow paths can stay inline until they have their own focused acceptance criteria.
Confidence: high
Scope-risk: moderate
Tested: deno test --no-check --allow-all src/workflow/executor/dag/index.test.ts
Tested: deno test --no-check --allow-all src/workflow/executor/dag/index.test.ts src/workflow/executor/dag-executor.test.ts src/workflow/executor/workflow-executor.test.ts src/workflow/__tests__/workflow.integration.test.ts
Tested: deno check src/workflow/executor/dag/index.ts src/workflow/executor/dag/index.test.ts src/workflow/executor/dag/map-node-strategy.ts src/workflow/executor/dag/loop-node-strategy.ts src/workflow/executor/dag/node-strategy-types.ts
Tested: git diff --check
Tested: deno task verify:quick
Related: #1984
