+++ src/workflow/executor/dag/index.ts
@@ -9,20 +9,16 @@
 import type {
   BranchNodeConfig,
   Checkpoint,
-  LoopExecutionContext,
-  LoopNodeConfig,
-  MapNodeConfig,
   NodeState,
   ParallelNodeConfig,
   SubWorkflowNodeConfig,
   WaitNodeConfig,
   WorkflowContext,
-  WorkflowDefinition,
   WorkflowNode,
   WorkflowNodeConfig,
   WorkflowRun,
 } from "../../types.ts";
-import { generateId, parseDuration } from "../../types.ts";
+import { generateId } from "../../types.ts";
 import { INVALID_ARGUMENT, NOT_SUPPORTED } from "#veryfront/errors";
 
 export type { DAGExecutionResult, DAGExecutorConfig, NodeExecutionResult } from "./types.ts";
@@ -33,8 +29,11 @@ import type {
   DAGExecutorInternalConfig,
   NodeExecutionResult,
 } from "./types.ts";
-import { deriveNodeStatus, shouldCheckpoint, sleep } from "./utils.ts";
+import { deriveNodeStatus, shouldCheckpoint } from "./utils.ts";
 import { buildGraph, getReadyNodes, hasCycle, updateInDegreesForCompletedNodes } from "./graph.ts";
+import { executeLoopNodeStrategy } from "./loop-node-strategy.ts";
+import { executeMapNodeStrategy } from "./map-node-strategy.ts";
+import type { ChildGraphExecutionOptions } from "./node-strategy-types.ts";
 
 export class DAGExecutor {
   private config: DAGExecutorInternalConfig;
@@ -181,7 +180,16 @@ export class DAGExecutor {
       case "parallel":
         return this.executeParallelNode(node, config, context, nodeStates);
       case "map":
-        return this.executeMapNode(node, config, context, nodeStates);
+        return executeMapNodeStrategy({
+          node,
+          config,
+          context,
+          nodeStates,
+          runtime: {
+            executeChildGraph: (nodes, run, options) => this.executeChildGraph(nodes, run, options),
+            onNodeComplete: this.config.onNodeComplete,
+          },
+        });
       case "branch":
         return this.executeBranchNode(node, config, context, nodeStates);
       case "wait":
@@ -189,7 +197,16 @@ export class DAGExecutor {
       case "subWorkflow":
         return this.executeSubWorkflowNode(node, config, context);
       case "loop":
-        return this.executeLoopNode(node, config, context, nodeStates);
+        return executeLoopNodeStrategy({
+          node,
+          config,
+          context,
+          nodeStates,
+          runtime: {
+            executeChildGraph: (nodes, run) => this.executeChildGraph(nodes, run),
+            onNodeComplete: this.config.onNodeComplete,
+          },
+        });
       default:
         throw INVALID_ARGUMENT.create({
           detail:
@@ -267,107 +284,6 @@ export class DAGExecutor {
       contextUpdates: result.context,
       waiting: result.waiting,
     };
-  }
-
-  private async executeMapNode(
-    node: WorkflowNode,
-    config: MapNodeConfig,
-    context: WorkflowContext,
-    nodeStates: Record<string, NodeState>,
-  ): Promise<NodeExecutionResult> {
-    const startTime = Date.now();
-
-    const items = typeof config.items === "function" ? await config.items(context) : config.items;
-
-    if (!Array.isArray(items)) {
-      throw INVALID_ARGUMENT.create({ detail: `Map node "${node.id}" items must be an array` });
-    }
-
-    if (items.length === 0) {
-      const state: NodeState = {
-        nodeId: node.id,
-        status: "completed",
-        output: [],
-        attempt: 1,
-        startedAt: new Date(startTime),
-        completedAt: new Date(),
-      };
-      return { state, contextUpdates: { [node.id]: [] }, waiting: false };
-    }
-
-    const isWorkflowDef = (p: unknown): p is WorkflowDefinition =>
-      typeof p === "object" && p !== null && "steps" in p;
-
-    const childNodes: WorkflowNode[] = items.map((item, i) => {
-      const childId = `${node.id}_${i}`;
-
-      if (isWorkflowDef(config.processor)) {
-        return {
-          id: childId,
-          config: {
-            type: "subWorkflow",
-            workflow: config.processor,
-            input: item,
-            retry: config.retry,
-            checkpoint: false,
-          },
-        };
-      }
-
-      const processorConfig: WorkflowNodeConfig = { ...(config.processor as WorkflowNode).config };
-
-      if (processorConfig.type === "step") {
-        processorConfig.input = item as Record<string, unknown>;
-      }
-
-      return { id: childId, config: processorConfig };
-    });
-
-    const originalConcurrency = this.config.maxConcurrency;
-    if (config.concurrency) {
-      this.config.maxConcurrency = config.concurrency;
-    }
-
-    try {
-      const result = await this.execute(childNodes, {
-        id: `${node.id}_map`,
-        workflowId: "",
-        status: "running",
-        input: context.input,
-        // Carry already-accumulated child states so completed children are
-        // skipped on resume instead of re-executing (H8).
-        nodeStates,
-        currentNodes: [],
-        context: { ...context },
-        checkpoints: [],
-        pendingApprovals: [],
-        createdAt: new Date(),
-      });
-
-      Object.assign(nodeStates, result.nodeStates);
-
-      const outputs = childNodes.map((child) => result.nodeStates[child.id]?.output);
-
-      const state: NodeState = {
-        nodeId: node.id,
-        status: deriveNodeStatus(result.completed, result.waiting),
-        output: outputs,
-        error: result.error,
-        attempt: 1,
-        startedAt: new Date(startTime),
-        completedAt: result.completed ? new Date() : undefined,
-      };
-
-      this.config.onNodeComplete?.(node.id, state);
-
-      return {
-        state,
-        contextUpdates: result.completed ? { [node.id]: outputs } : {},
-        waiting: result.waiting,
-      };
-    } finally {
-      this.config.maxConcurrency = originalConcurrency;
-    }
   }
 
   private async executeBranchNode(
@@ -526,172 +442,6 @@ export class DAGExecutor {
     };
   }
 
-  private async executeLoopNode(
-    node: WorkflowNode,
-    config: LoopNodeConfig,
-    context: WorkflowContext,
-    nodeStates: Record<string, NodeState>,
-  ): Promise<NodeExecutionResult> {
-    const startTime = Date.now();
-    const previousResults: unknown[] = [];
-    let iteration = 0;
-    let exitReason: "condition" | "maxIterations" | "error" = "condition";
-    let lastError: string | undefined;
-
-    const existingLoopState = context[`${node.id}_loop_state`] as
-      | {
-        iteration: number;
-        previousResults: unknown[];
-        iterationNodeStates?: Record<string, NodeState>;
-      }
-      | undefined;
-
-    // Child node states for the in-flight (resumed) iteration, so its already
-    // completed steps are not re-executed on resume (H9).
-    let resumeIterationNodeStates: Record<string, NodeState> | undefined;
-    let resumeIteration: number | undefined;
-
-    if (existingLoopState) {
-      iteration = existingLoopState.iteration;
-      previousResults.push(...existingLoopState.previousResults);
-      resumeIterationNodeStates = existingLoopState.iterationNodeStates;
-      resumeIteration = existingLoopState.iteration;
-    }
-
-    while (iteration < config.maxIterations) {
-      const loopContext: LoopExecutionContext = {
-        iteration,
-        totalIterations: iteration,
-        previousResults: [...previousResults],
-        isFirstIteration: iteration === 0,
-        isLastAllowedIteration: iteration === config.maxIterations - 1,
-      };
-
-      if (!(await config.while(context, loopContext))) {
-        exitReason = "condition";
-        break;
-      }
-
-      const steps = typeof config.steps === "function"
-        ? config.steps(context, loopContext)
-        : config.steps;
-
-      // On resume, rehydrate the in-flight iteration's child node states so its
-      // already-completed steps are skipped instead of re-executed (H9).
-      const iterationNodeStates = resumeIteration === iteration && resumeIterationNodeStates
-        ? { ...resumeIterationNodeStates }
-        : {};
-      // Only rehydrate once; subsequent iterations start fresh.
-      resumeIterationNodeStates = undefined;
-
-      const result = await this.execute(steps, {
-        id: `${node.id}_iter_${iteration}`,
-        workflowId: "",
-        status: "running",
-        input: context.input,
-        nodeStates: iterationNodeStates,
-        currentNodes: [],
-        context: { ...context, _loop: loopContext },
-        checkpoints: [],
-        pendingApprovals: [],
-        createdAt: new Date(),
-      });
-
-      if (result.waiting) {
-        Object.assign(nodeStates, result.nodeStates);
-
-        const state: NodeState = {
-          nodeId: node.id,
-          status: "running",
-          output: { iteration, waiting: true, previousResults },
-          attempt: 1,
-          startedAt: new Date(startTime),
-        };
-
-        return {
-          state,
-          contextUpdates: {
-            ...result.context,
-            [`${node.id}_loop_state`]: {
-              iteration,
-              previousResults,
-              // Persist the in-flight iteration's child states so completed
-              // steps are not re-executed when this iteration resumes (H9).
-              iterationNodeStates: result.nodeStates,
-            },
-          },
-          waiting: true,
-        };
-      }
-
-      if (result.error) {
-        lastError = result.error;
-        exitReason = "error";
-        break;
-      }
-
-      previousResults.push(result.context);
-      Object.assign(context, result.context);
-      Object.assign(nodeStates, result.nodeStates);
-
-      if (config.delay && iteration < config.maxIterations - 1) {
-        const delayMs = typeof config.delay === "number"
-          ? config.delay
-          : parseDuration(config.delay);
-        await sleep(delayMs);
-      }
-
-      iteration++;
-    }
-
-    if (iteration >= config.maxIterations && exitReason !== "condition") {
-      exitReason = "maxIterations";
-    }
-
-    const finalLoopContext: LoopExecutionContext = {
-      iteration,
-      totalIterations: iteration,
-      previousResults,
-      isFirstIteration: false,
-      isLastAllowedIteration: true,
-    };
-
-    let completionUpdates: Record<string, unknown> = {};
-    if (exitReason === "maxIterations" && config.onMaxIterations) {
-      completionUpdates = await config.onMaxIterations(context, finalLoopContext);
-    } else if (exitReason === "condition" && config.onComplete) {
-      completionUpdates = await config.onComplete(context, finalLoopContext);
-    }
-
-    const output = {
-      exitReason,
-      iterations: iteration,
-      previousResults,
-      ...completionUpdates,
-    };
-
-    const state: NodeState = {
-      nodeId: node.id,
-      status: exitReason === "error" ? "failed" : "completed",
-      output,
-      error: lastError,
-      attempt: 1,
-      startedAt: new Date(startTime),
-      completedAt: new Date(),
-    };
-
-    this.config.onNodeComplete?.(node.id, state);
-
-    return {
-      state,
-      contextUpdates: {
-        [node.id]: output,
-        ...completionUpdates,
-      },
-      waiting: false,
-    };
-  }
-
   private async checkpoint(
     runId: string,
     nodeId: string,
@@ -711,5 +461,23 @@ export class DAGExecutor {
     };
 
     await this.config.checkpointManager.save(runId, checkpoint);
+  }
+
+  private async executeChildGraph(
+    nodes: WorkflowNode[],
+    run: WorkflowRun,
+    options?: ChildGraphExecutionOptions,
+  ): Promise<DAGExecutionResult> {
+    if (!options?.maxConcurrency) {
+      return await this.execute(nodes, run);
+    }
+
+    const originalConcurrency = this.config.maxConcurrency;
+    this.config.maxConcurrency = options.maxConcurrency;
+    try {
+      return await this.execute(nodes, run);
+    } finally {
+      this.config.maxConcurrency = originalConcurrency;
+    }
   }
 }
